+++ bin/question/question.js
@@ -1,6 +1,20 @@
 const inquirer = require("inquirer");
 
 module.exports = {
+
+    async menuQuestion () {
+        const menuQuestion = [
+            {
+                type: 'list',
+                name: "menu",
+                message: "Que voulez-vous faire ? :",
+                choices: ['Vaccine', 'Tracks covid-19', 'Make attestation'],
+                default: "VACCINE"
+            },
+        ];
+        return inquirer.prompt(menuQuestion);
+    },
+    
     basicQuestion (question) {
         const basicQuestion = [
             {
